+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added enough code so that `addCollections()` does not error
 - Added enough working code so that collections are removable.
 - Added enough code so that some queries work
+- Added enough types that the TypeScript compiler is happy